Finish fixing energy and fluid capabilities

+++ src/main/java/fr/frinn/custommachinery/client/screen/creation/component/builder/FluidComponentBuilder.java
@@ -59,7 +59,7 @@ public class FluidComponentBuilder implements IMachineComponentBuilder<FluidMach
 
         @Override
         public Template makeTemplate() {
-            return new Template(this.id.getValue(), this.parseLong(this.capacity.getValue()), this.parseLong(this.maxInput.getValue()), this.parseLong(this.maxOutput.getValue()), Collections.emptyList(), false, this.mode.getValue(), this.mode.getValue().getBaseConfig(), this.unique.selected());
+            return new Template(this.id.getValue(), (int)this.parseLong(this.capacity.getValue()), (int)this.parseLong(this.maxInput.getValue()), (int)this.parseLong(this.maxOutput.getValue()), Collections.emptyList(), false, this.mode.getValue(), this.mode.getValue().getBaseConfig(), this.unique.selected());
         }
 
         @Override
